player title and more explores out of buttonhelper
