+++ src/main/java/ti4/helpers/ExploreHelper.java
@@ -6,11 +6,19 @@ import org.apache.commons.lang3.StringUtils;
 
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent;
 import net.dv8tion.jda.api.interactions.components.buttons.Button;
+import ti4.commands.cardsac.ACInfo;
+import ti4.commands.leaders.CommanderUnlockCheck;
+import ti4.commands.leaders.RefreshLeader;
+import ti4.commands.planet.PlanetRefresh;
 import ti4.commands.units.AddUnits;
 import ti4.generator.Mapper;
+import ti4.helpers.Units.UnitKey;
 import ti4.listeners.annotations.ButtonHandler;
 import ti4.map.Game;
+import ti4.map.Leader;
 import ti4.map.Player;
+import ti4.map.Tile;
+import ti4.map.UnitHolder;
 import ti4.message.MessageHelper;
 
 public class ExploreHelper {
@@ -42,7 +50,7 @@ public class ExploreHelper {
     @ButtonHandler("resolveVolatile_")
     public static void resolveVolatileFuelSource(String buttonID, Game game, Player player, ButtonInteractionEvent event) {
         String planetID = StringUtils.substringAfter(buttonID, "_");
-        String mechOrInfCheckMessage = ButtonHelper.mechOrInfCheck(planetID, game, player);
+        String mechOrInfCheckMessage = checkForMechOrRemoveInf(planetID, game, player);
         boolean failed = mechOrInfCheckMessage.contains("Please try again.");
 
         if (!failed) {
@@ -57,6 +65,139 @@ public class ExploreHelper {
             if (!game.isFowMode() && (event.getChannel() != game.getActionsChannel())) {
                 MessageHelper.sendMessageToChannel(player.getCorrectChannel(), player.getFactionEmoji() + " " + mechOrInfCheckMessage);
             }
+        }
+    }
+
+    @ButtonHandler("resolveExpedition_")
+    public static void resolveExpedition(String buttonID, Game game, Player player, ButtonInteractionEvent event) {
+        String message = "";
+        String planetName = buttonID.split("_")[1];
+        boolean failed = false;
+        message = message + checkForMechOrRemoveInf(planetName, game, player);
+        failed = message.contains("Please try again.");
+        if (!failed) {
+            PlanetRefresh.doAction(player, planetName, game);
+            planetName = Mapper.getPlanet(planetName) == null ? planetName : Mapper.getPlanet(planetName).getName();
+            message = message + "Readied " + planetName;
+            ButtonHelper.addReaction(event, false, false, message, "");
+            ButtonHelper.deleteMessage(event);
+        } else {
+            ButtonHelper.addReaction(event, false, false, message, "");
+        }
+    }
+
+    @ButtonHandler("resolveCoreMine_")
+    public static void resolveCoreMine(String buttonID, Game game, Player player, ButtonInteractionEvent event) {
+        String message = "";
+        String planetName = buttonID.split("_")[1];
+        boolean failed = false;
+        message = message + checkForMechOrRemoveInf(planetName, game, player);
+        failed = message.contains("Please try again.");
+        if (!failed) {
+            message = message + "Gained 1TG " + player.gainTG(1, true);
+            ButtonHelperAgents.resolveArtunoCheck(player, game, 1);
+        }
+        ButtonHelper.addReaction(event, false, false, message, "");
+        if (!failed) {
+            ButtonHelper.deleteMessage(event);
+            if (!game.isFowMode() && (event.getChannel() != game.getActionsChannel())) {
+                String pF = player.getFactionEmoji();
+                MessageHelper.sendMessageToChannel(player.getCorrectChannel(), pF + " " + message);
+            }
+        }
+    }
+
+    public static String checkForMechOrRemoveInf(String planetName, Game game, Player player) {
+        String message;
+        Tile tile = game.getTile(AliasHandler.resolveTile(planetName));
+        UnitHolder unitHolder = tile.getUnitHolders().get(planetName);
+        int numMechs = 0;
+        int numInf = 0;
+        String colorID = Mapper.getColorID(player.getColor());
+        UnitKey mechKey = Mapper.getUnitKey("mf", colorID);
+        UnitKey infKey = Mapper.getUnitKey("gf", colorID);
+        if (unitHolder.getUnits() != null) {
+
+            if (unitHolder.getUnits().get(mechKey) != null) {
+                numMechs = unitHolder.getUnits().get(mechKey);
+            }
+            if (unitHolder.getUnits().get(infKey) != null) {
+                numInf = unitHolder.getUnits().get(infKey);
+            }
+        }
+        if (numMechs > 0 || numInf > 0) {
+            if (numMechs > 0) {
+                message = "Planet had a mech. ";
+            } else {
+                message = "Planet did not have a mech. Removed 1 infantry (" + numInf + "->" + (numInf - 1) + "). ";
+                tile.removeUnit(planetName, infKey, 1);
+            }
+        } else {
+            message = "Planet did not have a mech or an infantry. Please try again.";
+        }
+        return message;
+    }
+
+    @ButtonHandler("ruins_")
+    public static void resolveWarForgeRuins(Game game, String buttonID, Player player, ButtonInteractionEvent event) {
+        String planet = buttonID.split("_")[1];
+        String mech = buttonID.split("_")[2];
+        String message = "";
+        boolean failed;
+        message = message + checkForMechOrRemoveInf(planet, game, player);
+        failed = message.contains("Please try again.");
+        if (!failed) {
+            if ("mech".equalsIgnoreCase(mech)) {
+                new AddUnits().unitParsing(event, player.getColor(), game.getTileFromPlanet(planet),
+                    "mech " + planet, game);
+                message = message + "Placed mech on" + Mapper.getPlanet(planet).getName();
+            } else {
+                new AddUnits().unitParsing(event, player.getColor(), game.getTileFromPlanet(planet),
+                    "2 infantry " + planet, game);
+                message = message + "Placed 2 infantry on" + Mapper.getPlanet(planet).getName();
+            }
+            ButtonHelper.addReaction(event, false, false, message, "");
+            ButtonHelper.deleteMessage(event);
+        } else {
+            ButtonHelper.addReaction(event, false, false, message, "");
+        }
+    }
+
+    @ButtonHandler("seedySpace_")
+    public static void resolveSeedySpace(Game game, String buttonID, Player player, ButtonInteractionEvent event) {
+        String planet = buttonID.split("_")[2];
+        String acOrAgent = buttonID.split("_")[1];
+        String message = "";
+        boolean failed;
+        message = message + checkForMechOrRemoveInf(planet, game, player);
+        failed = message.contains("Please try again.");
+        if (!failed) {
+            if ("ac".equalsIgnoreCase(acOrAgent)) {
+                if (player.hasAbility("scheming")) {
+                    game.drawActionCard(player.getUserID());
+                    game.drawActionCard(player.getUserID());
+                    message = player.getFactionEmoji() + " Drew 2 ACs with Scheming. Please discard 1 AC with the blue buttons.";
+                    MessageHelper.sendMessageToChannelWithButtons(player.getCardsInfoThread(),
+                        player.getRepresentation(true, true) + " use buttons to discard",
+                        ACInfo.getDiscardActionCardButtons(game, player, false));
+                } else {
+                    game.drawActionCard(player.getUserID());
+                    message = player.getFactionEmoji() + " Drew 1 AC";
+                    ACInfo.sendActionCardInfo(game, player, event);
+                }
+                CommanderUnlockCheck.checkPlayer(player, "yssaril");
+            } else {
+                Leader playerLeader = player.getLeader(acOrAgent).orElse(null);
+                if (playerLeader == null) {
+                    return;
+                }
+                RefreshLeader.refreshLeader(player, playerLeader, game);
+                message = message + " Refreshed " + Mapper.getLeader(acOrAgent).getName();
+            }
+            ButtonHelper.addReaction(event, false, false, message, "");
+            ButtonHelper.deleteMessage(event);
+        } else {
+            ButtonHelper.addReaction(event, false, false, message, "");
         }
     }
 }
